more conmtroller updates

+++ Controllers/DtpDatasetsApiController.cs
@@ -13,37 +13,20 @@ public class DtpDatasetsApiController : BaseApiController
     {
         _rmsService = rmsService ?? throw new ArgumentNullException(nameof(rmsService));
     }
-
-    /****************************************************************
-    * FETCH ALL datasets
-    ****************************************************************/
-
-    [HttpGet("data-transfers/datasets")]
-    [SwaggerOperation(Tags = new []{"Data transfer process datasets endpoint"})]
-    
-    public async Task<IActionResult> GetDtpDatasetList()
-    {
-        var datasets = await _rmsService.GetAllDtpDatasetsAsync();
-        if (datasets == null || datasets.Count == 0)
-        {
-            return Ok(NoAttributesResponse<DtpDataset>("No datasets were found."));
-        }   
-        return Ok(new ApiResponse<DtpDataset>()
-        {
-            Total = datasets.Count, StatusCode = Ok().StatusCode, Messages = null,
-            Data = datasets
-        });
-    }
     
     /****************************************************************
-    * FETCH a particular dateset record
+    * FETCH a particular dateset record, for a specified object
     ****************************************************************/
     
-    [HttpGet("data-transfers/datasets/{id:int}")]
+    [HttpGet("data-transfers/{sd_oid}/datasets/{id:int}")]
     [SwaggerOperation(Tags = new []{"Data transfer process datasets endpoint"})]
     
-    public async Task<IActionResult> GetDtpDataset(int id)
+    public async Task<IActionResult> GetDtpDataset(string sd_oid, int id)
     {
+        if (await _rmsService.ObjectDatasetDoesNotExistAsync(sd_oid, id))
+        {
+            return Ok(ErrorInActionResponse<ObjectInstance>("No dataset with that id found for specified object."));
+        }
         var dtpDataset = await _rmsService.GetDtpDatasetAsync(id);
         if (dtpDataset == null) 
         {
@@ -60,13 +43,13 @@ public class DtpDatasetsApiController : BaseApiController
     * CREATE a new dataset record, linked to a specified object
     ****************************************************************/
     
-    [HttpPost("data-transfers/datasets/{object_id}")]
+    [HttpPost("data-transfers/{sd_oid}/datasets")]
     [SwaggerOperation(Tags = new []{"Data transfer process datasets endpoint"})]
     
-    public async Task<IActionResult> CreateDtpDataset(string object_id, 
+    public async Task<IActionResult> CreateDtpDataset(string sd_oid, 
         [FromBody] DtpDataset dtpDatasetContent)
     {
-        dtpDatasetContent.ObjectId = object_id;
+        dtpDatasetContent.ObjectId = sd_oid;
         var dataset = await _rmsService.CreateDtpDatasetAsync(dtpDatasetContent);
         if (dataset == null)
         {
@@ -84,13 +67,16 @@ public class DtpDatasetsApiController : BaseApiController
     * UPDATE a specific dataset record details
     ****************************************************************/
     
-    [HttpPut("data-transfers/datasets/{id:int}")]
+    [HttpPut("data-transfers/{sd_oid}/datasets/{id:int}")]
     [SwaggerOperation(Tags = new []{"Data transfer process datasets endpoint"})]
     
-    public async Task<IActionResult> UpdateDtpDataset(int id, 
+    public async Task<IActionResult> UpdateDtpDataset(string sd_oid, int id, 
         [FromBody] DtpDataset dtpDatasetContent)
     {
- // check object and dataset exist
+        if (await _rmsService.ObjectDatasetDoesNotExistAsync(sd_oid, id))
+        {
+            return Ok(ErrorInActionResponse<ObjectInstance>("No dataset with that id found for specified object."));
+        }
         var updatedDataset = await _rmsService.UpdateDtpDatasetAsync(id, dtpDatasetContent);
         if (updatedDataset == null)
         {
@@ -107,12 +93,15 @@ public class DtpDatasetsApiController : BaseApiController
     * DELETE a specified dataset record
     ****************************************************************/
     
-    [HttpDelete("data-transfers/datasets/{id:int}")]
+    [HttpDelete("data-transfers/{sd_oid}/datasets/{id:int}")]
     [SwaggerOperation(Tags = new []{"Data transfer process datasets endpoint"})]
     
-    public async Task<IActionResult> DeleteDtpDataset(int id)
+    public async Task<IActionResult> DeleteDtpDataset(string sd_oid, int id)
     {
-        // check object and dataset exist
+        if (await _rmsService.ObjectDatasetDoesNotExistAsync(sd_oid, id))
+        {
+            return Ok(ErrorInActionResponse<ObjectInstance>("No dataset with that id found for specified object."));
+        }
         var count = await _rmsService.DeleteDtpDatasetAsync(id);
         return Ok(new ApiResponse<DtpDataset>()
         {
